analizeSentimen/requirements.txt
functions-framework==3.*
google-cloud-storage==2.13.0
scikit-learn==1.3.0
